+++ packages/afterburner/index.ts
@@ -1,4 +1,5 @@
 import * as appsync from './appsync/';
 import * as apigatewayv2 from './apigatewayv2/';
+import * as cognito from './cognito/';
 
-export { appsync, apigatewayv2 };
+export { appsync, apigatewayv2, cognito };