+++ packages/jet/flight/commands/common/types.ts
@@ -1,15 +1,34 @@
-export interface Stack {
+// Normal deployed stack info might not have jet, if it has no output files
+export interface DeployedStack {
+  jet?: string;
+}
+// Normal deployed stack info might not have jet, if it has no output files
+export interface ParsedDeployedStack extends Record<string, any> {
+  jet?: JetOutput;
+}
+
+export interface DeployedDevStack extends DeployedStack {
   jet: string;
 }
-export interface ParsedStack extends Record<string, any> {
+export interface ParsedDeployedDevStack extends ParsedDeployedStack {
   jet: JetOutput;
 }
 
+export function isDeployedDevStack(
+  stack: DeployedStack
+): stack is DeployedDevStack {
+  return stack.jet ? true : false;
+}
+export function isParsedDeployedDevStack(
+  stack: ParsedDeployedStack
+): stack is ParsedDeployedDevStack {
+  return stack.jet?.id ? true : false;
+}
 export interface JetOutput {
   id: string;
   functions: DeployedFunction[];
   assemblyOutDir: string;
-  writeValues: WriteValues[];
+  outputsFiles: OutputsFile[];
 }
 
 export interface SynthedFunction {
@@ -21,8 +40,15 @@ export interface DeployedFunction {
   name: string;
 }
 
-export interface WriteValues {
+export type OutputsFile = {
   path: string;
-  values: Record<string, string>;
-  format?: 'json' | 'json5' | 'env';
-}
+} & (
+  | {
+      contents: Record<string, unknown>;
+      format?: 'json' | 'json5';
+    }
+  | {
+      contents: Record<string, string>;
+      format: 'env';
+    }
+);