+++ packages/afterburner/CHANGELOG.md
@@ -1,12 +1,15 @@
-0.1.1 - September 29 2021
+# 0.2.0 - September 29 2021
+- Add cognito triggers api
+
+# 0.1.1 - September 29 2021
 - Fix publish of wrong folder
 
-0.1.0 - September 29 2021
+# 0.1.0 - September 29 2021
 - Add all missing appsync datasources, integrations, and authorizers
 - Rename imports to stick to aws naming convention 
 - Rename fns to functions
 - Add resolvers function to explicitly create resolvers, make it necessary for use for lambda datasource 
 
-0.0.2 - September 28 2021
+# 0.0.2 - September 28 2021
 - Remove some imports from index files to avoid unnecessary dependencies
 - Change id generation of datasources to avoid '-'